+++ performance/os-performance/README.md
@@ -52,3 +52,13 @@ os-performace -u admin -p secred -n 1000 -t connect
 
 This test opens 1000 autoupdate connections and shows, when they receive an
 update.
+
+
+## Create dummy users
+
+```
+os-performance -u admin -p secred -n 1000 create_users
+```
+
+This creates 1.000 users with the names dummy1, dummy2 etc and the password
+"pass".
