+++ performance/os-performance/main.go
@@ -5,19 +5,14 @@ package main
 import (
 	"bufio"
 	"context"
-	"crypto/tls"
 	"encoding/json"
 	"errors"
-	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
-	"net/http/cookiejar"
 	"os"
 	"os/signal"
-	"strings"
 	"sync"
 	"time"
 
@@ -63,6 +58,8 @@ func run(ctx context.Context, args []string) error {
 		err = runBrowser(ctx, cfg)
 	case testConnect:
 		err = runKeepOpen(ctx, cfg)
+	case testCreateUsers:
+		err = runCreateUsers(ctx, cfg)
 	default:
 		err = fmt.Errorf("unknown testCase")
 	}
@@ -220,117 +217,6 @@ func scannerNotify(line string) (string, error) {
 	return fmt.Sprintf("notify name %s", format.Name), nil
 }
 
-type incrementer interface {
-	Increment()
-}
-
-type client struct {
-	domain             string
-	hc                 *http.Client
-	username, password string
-	inc                incrementer
-}
-
-// newClient creates a client object. No requests are sent.
-func newClient(domain, username, password string, inc incrementer) (*client, error) {
-	c := &client{
-		domain:   "https://" + domain,
-		username: username,
-		password: password,
-		inc:      inc,
-	}
-
-	jar, err := cookiejar.New(nil)
-	if err != nil {
-		return nil, fmt.Errorf("creating cookie jar: %w", err)
-	}
-	c.hc = &http.Client{
-		Jar: jar,
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		},
-	}
-	return c, nil
-}
-
-// browser sends the request each browser tab sends.
-func (c *client) browser() error {
-	if err := c.login(); err != nil {
-		return fmt.Errorf("login client: %w", err)
-	}
-
-	if c.inc != nil {
-		c.inc.Increment()
-	}
-
-	var wg sync.WaitGroup
-
-	for _, path := range []string{
-		pathWhoami,
-		pathLogin,
-		pathServertime,
-		pathConstants,
-	} {
-		wg.Add(1)
-		go func(path string) {
-			defer wg.Done()
-
-			if c.inc != nil {
-				defer c.inc.Increment()
-			}
-
-			if err := c.get(context.Background(), path); err != nil {
-				log.Printf("Error get request to %s: %v", path, err)
-			}
-		}(path)
-	}
-
-	wg.Wait()
-	return nil
-}
-
-func (c *client) keepOpen(ctx context.Context, path string) (io.ReadCloser, error) {
-	req, err := http.NewRequestWithContext(ctx, "GET", c.domain+path, nil)
-	if err != nil {
-		return nil, fmt.Errorf("creating request: %w", err)
-	}
-
-	resp, err := checkStatus(c.hc.Do(req))
-	if err != nil {
-		return nil, fmt.Errorf("sending %s request: %w", path, err)
-	}
-	return resp.Body, nil
-}
-
-// login uses the username and password to login the client. Sets the returned
-// cookie for later requests.
-func (c *client) login() error {
-	url := c.domain + pathLogin
-	payload := fmt.Sprintf(`{"username": "%s", "password": "%s"}`, c.username, c.password)
-
-	resp, err := checkStatus(c.hc.Post(url, "application/json", strings.NewReader(payload)))
-	if err != nil {
-		return fmt.Errorf("sending login request: %w", err)
-	}
-	resp.Body.Close()
-	return nil
-}
-
-func (c *client) get(ctx context.Context, path string) error {
-	req, err := http.NewRequestWithContext(ctx, "GET", c.domain+path, nil)
-	if err != nil {
-		return fmt.Errorf("creating request: %w", err)
-
-	}
-
-	resp, err := checkStatus(c.hc.Do(req))
-	if err != nil {
-		return fmt.Errorf("sending %s request: %w", path, err)
-	}
-	resp.Body.Close()
-	return nil
-}
-
 func checkStatus(resp *http.Response, err error) (*http.Response, error) {
 	if err != nil {
 		return nil, fmt.Errorf("sending login request: %w", err)
@@ -346,50 +232,3 @@ func checkStatus(resp *http.Response, err error) (*http.Response, error) {
 	}
 	return resp, nil
 }
-
-// Config contains all settings that can be changed with command line options.
-type Config struct {
-	testCase    int
-	clientCount int
-	domain      string
-	username    string
-	passwort    string
-	url         string
-}
-
-func loadConfig(args []string) (*Config, error) {
-	cfg := new(Config)
-	f := flag.NewFlagSet("args", flag.ExitOnError)
-
-	f.IntVar(&cfg.clientCount, "n", 10, "number of connections to use")
-	f.StringVar(&cfg.domain, "d", "localhost:8000", "host and port of the server to test")
-	f.StringVar(&cfg.username, "u", "admin", "username to use for login")
-	f.StringVar(&cfg.passwort, "p", "admin", "password to use for login")
-	f.StringVar(&cfg.url, "url", "autoupdate", "only for the `connect` test. autoupdate, projector or notify")
-
-	test := f.String("t", "", "testcase [browser,connect]")
-
-	if err := f.Parse(args); err != nil {
-		return nil, fmt.Errorf("parsing flags: %w", err)
-	}
-
-	if len(flag.Args()) > 0 {
-		return nil, fmt.Errorf("invalid arguments: %s", strings.Join(flag.Args(), " "))
-	}
-
-	switch *test {
-	case "browser":
-		cfg.testCase = testBrowser
-	case "connect":
-		cfg.testCase = testConnect
-	default:
-		return nil, fmt.Errorf("invalid testcase %s", *test)
-	}
-
-	return cfg, nil
-}
-
-const (
-	testBrowser = iota
-	testConnect = iota
-)
